select dropdown default store

+++ convex/categories.ts
@@ -5,14 +5,14 @@ import { Id } from "./_generated/dataModel";
 
 export const get = query({
   handler: async (ctx) => {
-    const identity = await authorizeUser(ctx, "No Auth: add products");
+    const identity = await authorizeUser(ctx, "No Auth: view categories");
     const user = await ctx.db
       .query("users")
       .withIndex("by_token", (q) => {
         return q.eq("tokenIdentifier", identity.tokenIdentifier);
       })
       .unique();
-    if (!user) return;
+    if (!user) throw new Error("No User Found");
     if (user.default_store) {
       const categories = await ctx.db
         .query("categories")
@@ -22,7 +22,7 @@ export const get = query({
         .collect();
       return categories;
     } else {
-      return "no category";
+      return "no_default_store";
     }
   },
 });
@@ -32,18 +32,18 @@ export const add = mutation({
     name: v.string(),
   },
   handler: async (ctx, args) => {
-    const identity = await authorizeUser(ctx, "No Auth: add products");
+    const identity = await authorizeUser(ctx, "No Auth: add categories");
     const user = await ctx.db
       .query("users")
       .withIndex("by_token", (q) => {
         return q.eq("tokenIdentifier", identity.tokenIdentifier);
       })
       .unique();
-
-    if (user?.default_store) {
+    if (!user) throw new Error("No User Found");
+    if (user.default_store) {
       await ctx.db.insert("categories", {
         name: args.name,
-        store_id: user?.default_store,
+        store_id: user.default_store,
       });
     }
   },
@@ -55,6 +55,18 @@ export const patch = mutation({
     name: v.string(),
   },
   handler: async (ctx, args) => {
+    const identity = await authorizeUser(ctx, "No Auth: patch categories");
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) => {
+        return q.eq("tokenIdentifier", identity.tokenIdentifier);
+      })
+      .unique();
+    if (!user) throw new Error("No User Found");
+    const category = await ctx.db.get(args.id);
+    if (category?.store_id !== user.default_store) {
+      throw new Error("default_store Don't Match: patch category");
+    }
     await ctx.db.patch(args.id, { name: args.name });
   },
 });
@@ -64,6 +76,18 @@ export const remove = mutation({
     id: v.id("categories"),
   },
   handler: async (ctx, args) => {
+    const identity = await authorizeUser(ctx, "No Auth: delete category");
+    const user = await ctx.db
+      .query("users")
+      .withIndex("by_token", (q) => {
+        return q.eq("tokenIdentifier", identity.tokenIdentifier);
+      })
+      .unique();
+    if (!user) throw new Error("No User Found");
+    const category = await ctx.db.get(args.id);
+    if (category?.store_id !== user.default_store) {
+      throw new Error("default_store Don't Match: delete category");
+    }
     await ctx.db.delete(args.id);
   },
 });
